Move test helper for ProcessingSettings to core module
Besides of one usage in formatting it is only used by the core module.
Custom rule sets (formatting is treated as one) should not depend on core.
