+++ detekt-core/src/test/kotlin/io/gitlab/arturbosch/detekt/core/DetektSpec.kt
@@ -1,6 +1,5 @@
 package io.gitlab.arturbosch.detekt.core
 
-import io.gitlab.arturbosch.detekt.test.createProcessingSettings
 import org.assertj.core.api.Assertions.assertThat
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
