+++ detekt-core/src/test/resources/configs/disabled-processors.yml
